% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{lbvem}
\alias{lbvem}
\title{co-clustering based on gaussian latent block}
\usage{
lbvem(X, g, m, n_init = 4, niter = 4)
}
\arguments{
\item{X}{Data matrix}

\item{g}{number of row clusters}

\item{m}{number of col clusters}

\item{n_init}{number of initialization for VEM and kmeans.}

\item{niter}{number of iterations for the co clustering algorithm}
}
\value{
list with model outputs
ICL: ICL value
Z_tild: a posteriori probabilities for row clusters
W_tild: a posteriori provabilities for col clusters
pi_k: proportions of row clusters
rho_l: proportions of col clusters
mu: co clusters mean matrix
sigma: co clusters variance matrix
}
\description{
This function performs a co-clustering by using a gaussian latent block model.
A gaussian distribution is used, and the parameters inference is realized
with latent block variational expectation-maximisation algorithm.
}
\examples{
#library(blockcluster)
#data(gaussiandata)
#model <- lbvem(gaussiandata, g=3, m=4, niter=5)
}
