% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotcoclust}
\alias{plotcoclust}
\title{plot co clustered data}
\usage{
plotcoclust(X, model)
}
\arguments{
\item{X}{Data matrix}

\item{model}{lbvem object}
}
\value{
nothing
}
\description{
This function plots data from a lbvem object. Comparing side by side raw data and co clustered data
}
\examples{
#library(blockcluster)
#data(gaussiandata)
#model <- lbvem(gaussiandata, g=3, m=4, niter=5)
#plotcoclust(gaussiandata, model)
}
